% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accelerometer_data}
\alias{accelerometer_data}
\title{Accelerometer sensor measurements}
\format{An 990 x 4 data frame with column names \code{t}, \code{x},
\code{y}, \code{z} containing accelerometer measurements. \code{t} is
the timestamp of each measurement, and \code{x}, \code{y} and \code{z}
are linear acceleration measurements along the respective axis.}
\usage{
accelerometer_data
}
\description{
Accelerometer sensor measurements
}
\keyword{datasets}
