% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{accelerometer_features}
\alias{accelerometer_features}
\title{Extract accelerometer features}
\usage{
accelerometer_features(sensor_data, window_length = 256, overlap = 0.5,
  bandpass_frequency_low = 1, bandpass_frequency_high = 25)
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z containing 
accelerometer measurements.}

\item{window_length}{Length of sliding windows.}

\item{overlap}{Window overlap.}

\item{bandpass_frequency_low}{Lower bound on frequency in Hz.}

\item{bandpass_frequency_high}{Upper bound on frequency in Hz.}
}
\value{
Accelerometer features.
}
\description{
Extract features summarizing time domain, frequency domain, and frequency energy
from accelerometer measurements.
}
