% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{accelerometer_features_}
\alias{accelerometer_features_}
\title{Extract accelerometer features}
\usage{
accelerometer_features_(sensor_data, transform, extract, groups,
  extract_on = c("acceleration", "jerk", "velocity", "displacement"))
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z containing 
accelerometer measurements.}

\item{transform}{A function which accepts \code{sensor_data} as input
and outputs a dataframe.}

\item{extract}{A list of functions to be applied to each of the columns
in \code{extract_on} from the output of \code{transform}.}

\item{groups}{A list of column names to group upon when calculating statistics.}

\item{extract_on}{A list of column names to calculate statistics from.}
}
\value{
Accelerometer features.
}
\description{
Extract accelerometer features. This function is not normally called
directly. See \code{accelerometer_features}.
}
