% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_drift}
\alias{calculate_drift}
\title{Calculate the drift given x and y}
\usage{
calculate_drift(x, y)
}
\arguments{
\item{x}{A vecor containing x co-ordinates (same length as that of y)}

\item{y}{A vecor containing y co-ordinates (same length as that of x)}
}
\value{
Drift vector which is sqrt(dx^2 + dy^2)
}
\description{
Calculate the drift given x and y
}
