% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tapping_features.R
\name{clean_tapped_button_none}
\alias{clean_tapped_button_none}
\title{Remove duplicates in the given dataframe tap_data which have the buttonid parameter as 'TappedButtonNone'}
\usage{
clean_tapped_button_none(tap_data)
}
\arguments{
\item{tap_data}{A n x 4 data frame with column names as t, x, y, buttonid containing 
tapping measurements. HEre n is the total number of tapping measurements, t is the time 
of measurement when the screen press was recorded, x and y are the coordinates of the screen 
touch in the smartphone, and buttonid can be one of the following values c('TappedButtonLeft',
'TappedButtonRight','TappedButtonNone') indicating a tap that has been classified as to the left,
right or neither of those places on the screen}
}
\value{
A dataframe with duplicates corresponding to buttonid == 'TappedButtonNone' removed
}
\description{
Remove duplicates in the given dataframe tap_data which have the buttonid parameter as 'TappedButtonNone'
}
