% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tapping_features.R
\name{CleanTappedButtonNone}
\alias{CleanTappedButtonNone}
\title{Remove duplicates in the given dataframe tap_data which have the buttonid parameter as 'TappedButtonNone'}
\usage{
CleanTappedButtonNone(tap_data)
}
\arguments{
\item{tap_data}{A dataframe with colums t,x,y and buttonid}
}
\value{
A dataframe with duplicates corresponding to buttonid == 'TappedButtonNone' removed
}
\description{
Remove duplicates in the given dataframe tap_data which have the buttonid parameter as 'TappedButtonNone'
}
