% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_var}
\alias{coef_var}
\title{Calculate the Coefficient of Variation (coef_var) for a given sequence}
\usage{
coef_var(x)
}
\arguments{
\item{x}{A numeric vector x whose Coefficient of Variation needs to be calculated}
}
\value{
A numeric value that is representative of the Coefficient of Variation
}
\description{
Calculate the Coefficient of Variation (coef_var) for a given sequence
}
