% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{createTappingFeaturesErrorResult}
\alias{createTappingFeaturesErrorResult}
\title{Create a tapping feature data frame NAs as features and an appropriate error message given as input}
\usage{
createTappingFeaturesErrorResult(error_message, featNames = c("meanTapInter",
  "medianTapInter", "iqrTapInter", "minTapInter", "maxTapInter", "skewTapInter",
  "kurTapInter", "sdTapInter", "madTapInter", "cvTapInter", "rangeTapInter",
  "tkeoTapInter", "ar1TapInter", "ar2TapInter", "fatigue10TapInter",
  "fatigue25TapInter", "fatigue50TapInter", "meanDriftLeft", "medianDriftLeft",
  "iqrDriftLeft", "minDriftLeft", "maxDriftLeft", "skewDriftLeft",
  "kurDriftLeft", "sdDriftLeft", "madDriftLeft", "cvDriftLeft",
  "rangeDriftLeft", "meanDriftRight", "medianDriftRight", "iqrDriftRight",     
  "minDriftRight", "maxDriftRight", "skewDriftRight", "kurDriftRight",
  "sdDriftRight", "madDriftRight", "cvDriftRight", "rangeDriftRight",
  "numberTaps", "buttonNoneFreq", "corXY"))
}
\arguments{
\item{error_message}{A text error message which would be reflected in the feature output}
}
\value{
A dataframe with NAs as feature values and the appropriate error message
}
\description{
Create a tapping feature data frame NAs as features and an appropriate error message given as input
}
