% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{derivative}
\alias{derivative}
\title{Take the derivative of a vector v}
\usage{
derivative(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Take the derivative of a vector v by calculating the difference
between component x_i and x_(i-1).
}
