% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detrend}
\alias{detrend}
\title{Detrend time series data}
\usage{
detrend(time, acceleration)
}
\arguments{
\item{time}{Numeric vector containing the timestamp values.}

\item{acceleration}{Numeric vector containing the acceleration values.}
}
\value{
Numeric vector with detrended acceleration values.
}
\description{
Detrend time series data
}
