% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{displacement}
\alias{displacement}
\title{Calculate displacement}
\usage{
displacement(acceleration, sampling_rate)
}
\arguments{
\item{acceleration}{An acceleration vector.}

\item{sampling_rate}{Sampling rate of the acceleration vector.}
}
\value{
Displacement vector.
}
\description{
Calculate displacement
}
