% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_features}
\alias{extract_features}
\title{Extract features from a column}
\usage{
extract_features(x, col, funs)
}
\arguments{
\item{x}{A grouped data frame with column \code{col}.}

\item{col}{The name of the column in \code{x} to pass to each
function in \code{funs}.}

\item{funs}{A list of functions that each accept a single vector as input.}
}
\value{
a data frame with the original grouped columns and
other feature columns.
}
\description{
Apply each of the functions in \code{funs} to the column
\code{col} in the grouped data frame \code{x}. Each of the functions in
\code{funs} must accept a single vector as input and output
a data frame.
}
