% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fatigue}
\alias{fatigue}
\title{Calculate the fatigue given a vector x}
\usage{
fatigue(x)
}
\arguments{
\item{x}{A numeric vector containing inter tap intevals}
}
\value{
A list containing fatigue10, fatigue25, fatigue50 where 
fatigueX is the difference in the mean values of the
first X percent of input x and last X percent of input x
}
\description{
Calculate the fatigue given a vector x
}
