% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtractionFunctions.R
\name{frequency_domain_energy}
\alias{frequency_domain_energy}
\title{Returns energy for each 0.5Hz band in the frequency spectrum.}
\usage{
frequency_domain_energy(values, samplig_rate)
}
\arguments{
\item{values}{A numeric vector.}

\item{sampling_rate}{Sampling_rate of \code{values}. If NA it uses default 
sampling rate of 100Hz.}
}
\value{
A features data frame of dimension 1 x 48, each representing energy within 
a 0.5Hz band.
}
\description{
\code{frequency_domain_energy()} is a convinient feature 
extraction function that converts a given time series in to frequency spectrum
and computes energy in each 0.5Hz band from 0 to 25 Hz.
}
\examples{
library(mhealthtools)

data(accelerometer_data)

fr.ftrs = frequency_domain_energy(accelerometer_data$x, sampling_rate = 100.122)
}
\author{
Thanneer Malai Perumal, Phil Snyder
}
