% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balance_features.R
\name{get_balance_features}
\alias{get_balance_features}
\title{Preprocess and extract interpretable features from balance assay.}
\usage{
get_balance_features(accelerometer_data = NULL, gyroscope_data = NULL,
  gravity_data = NULL, time_filter = NULL, detrend = F,
  frequency_filter = NULL, IMF = 1, window_length = NULL,
  window_overlap = NULL, derived_kinematics = F, funs = NULL,
  models = NULL)
}
\arguments{
\item{accelerometer_data}{An \code{n} x 4 data frame with columns \code{t}, \code{x},
\code{y}, \code{z} containing accelerometer measurements. Here \code{n} is the
total number of measurements, \code{t} is the timestamp of each measurement, and
\code{x}, \code{y} and \code{z} are linear acceleration measurements.}

\item{gyroscope_data}{An \code{n} x 4 data frame with columns \code{t}, \code{x},
\code{y}, \code{z} containing gyroscope measurements. Here \code{n} is the
total number of measurements, \code{t} is the timestamp of each measurement, and
\code{x}, \code{y} and \code{z} are linear velocity measurements.}

\item{gravity_data}{An \code{n} x 4 data frame with columns \code{t}, \code{x},
\code{y}, \code{z} containing gravity measurements. Here \code{n} is the
total number of measurements, \code{t} is the timestamp of each measurement, and
\code{x}, \code{y} and \code{z} are linear gravity measurements.}

\item{time_filter}{A length 2 numeric vector specifying the time range 
of measurements to use during preprocessing and feature extraction after
normalizing the first timestamp to 0. A \code{NULL} value means do not 
filter any measurements.}

\item{detrend}{A logical value indicating whether to detrend the signal. 
By default the value is FALSE.}

\item{frequency_filter}{A length 2 numeric vector specifying the frequency range
of the signal (in hertz) to use during preprocessing and feature extraction.
A \code{NULL} value means do not filter frequencies.}

\item{IMF}{The number of IMFs used during an empirical mode decomposition (EMD)
transformation. The default value of 1 means do not apply EMD to the signal.}

\item{window_length}{A numerical value representing the length (in number of samples)
of the sliding window used during the windowing transformation. Both 
\code{window_length} and \code{window_overlap} must be set for the windowing 
transformation to be applied.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap during a windowing transformation.
Both \code{window_length} and \code{window_overlap} must be set for the
windowing transformation to be applied.}

\item{derived_kinematics}{A logical value specifying whether to add derived
kinematic measurements (\code{displacement}, \code{velocity} or \code{acceleration},
and \code{jerk}) to \code{accelerometer_data} and \code{gyroscope_data} after
the transform defined by the above parameters has been applied to the raw
sensor measurements.}

\item{funs}{A function or list of functions that each accept a single numeric
vector as input. Each function should return a dataframe of features
(normally a single-row datafame, with column names as feature names).
The input vectors will be the axial measurements from \code{accelerometer_data}
and/or \code{gyroscope_data} after the transform defined by the above parameters
has been applied.If no argument is supplied to either \code{funs} or \code{models},
a default set of feature extraction functions (as described in
\code{default_kinematic_features}) will be supplied for this parameter.}

\item{models}{A function or list of functions that each accept
\code{sensor_data} as input after the transform defined by the above
parameters has been applied and returns features. Useful for functions
which compute individual features using multiple input variables.}
}
\value{
A list. The outputs from \code{funs} will
be stored under \code{$extracted_features} and the outputs from \code{models}
will be stored under \code{$model_features}. If there is an error 
during the transform process, an error dataframe will be stored under
\code{$error}. If gravity_data is passed and window_length and 
window_overlap are set, phone rotation information will be stored
under \code{$outlier_windows}.
}
\description{
A convenience wrapper for extracting interpretable features from the
balance assay measured using smartphone raw accelerometer and gyroscope
sensors. The balance assay entails participants standing still for 30 seconds
with the mobile device in their pocket or in a bag.
}
\examples{
balance_features <- get_balance_features(accelerometer_data, gyroscope_data)

balance_features <- get_balance_features(
  accelerometer_data,
  gyroscope_data,
  time_filter = c(2,4),
  detrend = TRUE,
  frequency_filter = c(0.5,25),
  window_length = 256,
  window_overlap = 0.2,
  derived_kinematics = TRUE)

balance_features <- get_balance_features(
  accelerometer_data,
  gyroscope_data,
  funs = list(time_domain_summary))
  
}
\seealso{
\code{\link{balance_data}}
}
\author{
Thanneer Malai Perumal, Meghasyam Tummalacherla, Phil Snyder
}
