% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{get_filtered_signal}
\alias{get_filtered_signal}
\title{Bandpass and sorted mean filter the given signal}
\usage{
get_filtered_signal(x, sampling_rate, mean_filter_order = 33,
  bandpass_order = 128, frequency_range = c(2, 25))
}
\arguments{
\item{x}{A time series numeric data}

\item{sampling_rate}{The sampling rate (fs) of the time series data}

\item{mean_filter_order}{Length of the sorted mean filter window}

\item{bandpass_order}{Order (length) of the bandpass filter to be used for filtering}

\item{frequency_range}{Frequency range in Hz for the bandpass filter parameters}
}
\value{
The filtered time series data
}
\description{
Bandpass and sorted mean filter the given signal
}
