% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{get_hr_from_time_series}
\alias{get_hr_from_time_series}
\title{Given a processed time series find its period using autocorrelation
and then convert it to heart rate (bpm)}
\usage{
get_hr_from_time_series(x, sampling_rate, method = "acf", min_hr = 40,
  max_hr = 200)
}
\arguments{
\item{x}{A time series numeric data}

\item{sampling_rate}{The sampling rate (fs) of the time series data}

\item{min_hr}{Minimum expected heart rate}

\item{max_hr}{Maximum expected heart rate}
}
\value{
A named vector containing heart rate and the confidence of the result
}
\description{
Given a processed time series find its period using autocorrelation
and then convert it to heart rate (bpm)
}
