% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sampling_rate}
\alias{get_sampling_rate}
\title{Calculate the sampling rate.}
\usage{
get_sampling_rate(sensor_data)
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z.}
}
\value{
The sampling rate (number of samples taken per second on average).
}
\description{
Calculate the sampling rate.
}
