% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tapping_features.R
\name{get_tapping_features}
\alias{get_tapping_features}
\title{Extract tapping features from raw tapping data.}
\usage{
get_tapping_features(tap_data, depress_threshold = 20,
  remove_duplicates = TRUE)
}
\arguments{
\item{tap_data}{A data frame with columns t, x, y, buttonid containing 
tapping measurements. buttonid can be from 
c('TappedButtonLeft','TappedButtonRight','TappedButtonNone') 
indicating a tap that has been classified as to the left, right 
or neither of those places on the screen}

\item{depress_threshold}{A numeric value indicating the threshold between 
taps}

\item{remove_duplicates}{A logical value indicating whether duplicates in taps
considered as None (i.e buttonid is neither left nor right) need to be removed.}
}
\value{
Tapping features
}
\description{
Extract tapping features from raw tapping data.
}
\author{
Elias Chaibub Neto, Meghasyam Tummalacherla, Phil Snyder
}
