% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTremorFeatures.R
\name{get_tremor_features}
\alias{get_tremor_features}
\title{Extract tremor features from raw accelerometer and gyroscope data.}
\usage{
get_tremor_features(accelerometer_data, gyrometer_data, gravity_data = NA,
  window_length = 256, frequency_range = c(1, 25), overlap = 0.5)
}
\arguments{
\item{accelerometer_data}{A data frame with columns t, x, y, z containing 
accelerometer measurements.}

\item{gyrometer_data}{A data frame with columns t, x, y, z containing 
gyrometer measurements.}

\item{gravity_data}{A data frame with columns t, x, y, z containing 
gravity sensor measurements.}

\item{window_length}{Length of sliding windows.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{overlap}{Window overlap.}
}
\value{
Tremor features indexed by axis and window.
}
\description{
Extract tremor features from raw accelerometer and gyroscope data.
}
\author{
Thanneer Malai Perumal, Meghasyam Tummalacherla, Phil Snyder
}
