% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_walk_features.R
\name{get_walk_features}
\alias{get_walk_features}
\title{Wrapper functionality to preprocess and extract interprettable features from 
walk assay measured using smartphone raw accelerometer and gyroscpe sensors.}
\usage{
get_walk_features(accelerometer_data, gyroscope_data,
  gravity_data = NULL, funs = NULL, models = NULL,
  window_length = 256, time_range = c(1, 9), frequency_range = c(1,
  25), overlap = 0.5, max_imf = 4)
}
\arguments{
\item{accelerometer_data}{A n x 4 data frame with column names as t, x, y, z 
containing accelerometer measurements. Here n is the total number of measurements, 
t is the time of measurement, x, y and z are the linear acceleration excluding 
gravity in their respective coordinates of the smartphone tri-axial accelerometers. 
Here x, y, and z are wrt phones axis as reference}

\item{gyroscope_data}{A n x 4 data frame with column names as t, x, y, z containing 
gyroscope measurements. Here n is the total number of measurements, t is the time 
of measurement, x, y and z are the angular velocity excluding gravity in their 
respective coordinates of the smartphone gyroscopes. Here x, y, and z are wrt phones 
axis as reference}

\item{gravity_data}{A n x 4 data frame with column names as t, x, y, z containing 
gravity estimations. Here n is the total number of measurements, t is the time of 
measurement, x, y and z are the linear acceleration measuring gravity components in
their respective coordinates of the smartphone gyroscopes. Here x, y, and z are wrt 
phones axis as reference}

\item{funs}{Feature extraction functions that accept a single time-series vector as input. 
Check extending vignette}

\item{models}{A function which accepts as input a dataframe with columns
axis, window, IMF, jerk, acceleration, velocity, displacement and
outputs features. Useful for models which compute individual statistics
using multiple input variables.}

\item{window_length}{Length of sliding windows for bandpass filter
and windowing transformation.}

\item{time_range}{Timestamp range to use.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{overlap}{Window overlap for the windowing transformation.}

\item{max_imf}{Number of intrinsic mode functions to use for 
empirical mode decomposition.}
}
\value{
A list of feature dataframes. The outputs from \code{funs} will
be stored under \code{$extracted_features} and the outputs from \code{models}
will be stored under \code{$model_features}.'
}
\description{
Wrapper functionality to preprocess and extract interprettable features from 
walk assay measured using smartphone raw accelerometer and gyroscpe sensors.
}
\author{
Thanneer Malai Perumal, Meghasyam Tummalacherla, Phil Snyder
}
