% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComparativeTremorFeatures.R
\name{getComparativeTremorFeatures}
\alias{getComparativeTremorFeatures}
\title{extracts features from tremor task handInLap or handAtShoulderLength accelerometer JSON data file by comparing left against right}
\usage{
getComparativeTremorFeatures(tremorRefJsonFileLoc, tremorAgtJsonFileLoc,
  windowLen = 256, freqRange = c(1, 25), ovlp = 0.5)
}
\arguments{
\item{tremorRefJsonFileLoc}{and tremorAgtJsonFileLoc path to tremor accelerometer json file}
}
\value{
data frame of tremor features
}
\description{
extracts features from tremor task handInLap or handAtShoulderLength accelerometer JSON data file by comparing left against right
}
