% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEWTspectrum}
\alias{getEWTspectrum}
\title{Get EWT spectrum}
\usage{
getEWTspectrum(spect, npeaks = 3, fractionMinPeakHeight = 0.1,
  minPeakDistance = 1, sampling_rate = 100)
}
\arguments{
\item{spect}{FFT spectrum as a two dimensional data frame with columns
names as freq and pdf respectively n.freq x 2.}

\item{npeaks}{Number of peaks to be captured.}

\item{fractionMinPeakHeight}{Minimum height (relative to maximum peak height)
a peak has to have. Specified as fraction between 0 and 1.}

\item{minPeakDistance}{The minimum distance (in indices) peaks.
have to have to be counted.}

\item{sampling_rate}{Sampling rate of the signal (by default it is 100 Hz).}
}
\value{
Emprical wavelet transformed spectrum of dimension n.freq x (npeaks + 1).
}
\description{
Given the spectrum of a time series vector this function will return its 
Empirical Wavelet Transformed spectrum.
}
