% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHR.R
\name{getHR}
\alias{getHR}
\title{extracts hr for each color band from avg pixel value per frame of video (processed hr) JSON data file}
\usage{
getHR(dat, windowLen = 10, freqRange = c(1, 25), bpforder = 128)
}
\value{
list containing hr and confidence of the estimate for each color (red, green, blue)
}
\description{
extracts hr for each color band from avg pixel value per frame of video (processed hr) JSON data file
}
\author{
Meghasyam Tummalacherla
}
