% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHrFromJson.R
\name{getHrFromJson}
\alias{getHrFromJson}
\title{extracts hr for each color band from avg pixel value per frame of video (processed hr) JSON data file}
\usage{
getHrFromJson(hrJsonFileLoc = NA, windowLen = 10, freqRange = c(1, 25),
  bpforder = 128)
}
\arguments{
\item{hrJsonFileLoc}{path to hr json file}
}
\value{
list containing hr and confidence of the estimate for each color (red, green, blue)
}
\description{
extracts hr for each color band from avg pixel value per frame of video (processed hr) JSON data file
}
\author{
Meghasyam Tummalacherla
}
