% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{getHrFromTimeSeries}
\alias{getHrFromTimeSeries}
\title{Given a processed time series find its period using autocorrelation and then convert it to heart rate (bpm)}
\usage{
getHrFromTimeSeries(x, samplingRate, minHR = 40, maxHR = 200)
}
\arguments{
\item{x}{A time series numeric data}

\item{samplingRate}{The sampling rate (fs) of the time series data}

\item{minHR}{Minimum expected heart rate}

\item{maxHR}{Maximum expected heart rate}
}
\value{
A named vector containing heart rate and the confidence of the result
}
\description{
Given a processed time series find its period using autocorrelation and then convert it to heart rate (bpm)
}
