% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoryModule.R
\name{getMemoryGameFeatures}
\alias{getMemoryGameFeatures}
\title{extracts memory game features from mPower game JSON data file}
\usage{
getMemoryGameFeatures(game_json_file)
}
\arguments{
\item{game_json_file}{path to game records json file}
}
\value{
data frame of memory game features
}
\description{
extracts memory game features from mPower game JSON data file
}
\examples{
library(synapseClient)
synapseLogin()
memoryTable = synTableQuery("SELECT * FROM syn5713115")
memoryTable = memoryTable@values
sampleRow = rownames(memoryTable)[1]
sample_gameRecord_File <- synDownloadTableFile('syn5713115', sampleRow,
"MemoryGameResults.json.MemoryGameGameRecords")
getMemoryGameFeatures(sample_gameRecord_File)
}
