% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingModule.R
\name{getPedometerFeatures}
\alias{getPedometerFeatures}
\title{extracts pedometer features from walking pedometer JSON data file}
\usage{
getPedometerFeatures(pedo_json_file)
}
\arguments{
\item{pedo_json_file}{path to pedometerjson file}
}
\value{
data frame of pedometer features
}
\description{
extracts pedometer features from walking pedometer JSON data file
}
\examples{
library(synapseClient)
synapseLogin()
sample_walking_pedometer_json_file <-'syn7315780'
pedometerJsonFile <- synGet(sample_walking_pedometer_json_file)@filePath
getPedometerFeatures(pedometerJsonFile)
}
