% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restModule.R
\name{getRestFeatures}
\alias{getRestFeatures}
\title{extracts features from accelerometer data file}
\usage{
getRestFeatures(restAccel_json_file)
}
\arguments{
\item{restAccel_json_file}{path to accelerometer json file}
}
\value{
data frame of rest features
}
\description{
extracts features from accelerometer data file
}

