% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpectrum}
\alias{getSpectrum}
\title{Get AR spectrum}
\usage{
getSpectrum(values, sampling_rate = 100, nfreq = 500)
}
\arguments{
\item{values}{A timeseries vector.}

\item{sampling_rate}{Sampling rate of the signal (by default it is 100 Hz).}

\item{nfreq}{Number of frequecy points to be interpolated.}
}
\value{
An AR spectrum.
}
\description{
Given a numeric vector, this function will return a spectrum
with all pole AR model.
}
