% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tappingModule.R
\name{getTappingFeatures}
\alias{getTappingFeatures}
\title{extracts tapping features from Tapping JSON data file}
\usage{
getTappingFeatures(tappingJsonFile, depressThr = 20)
}
\arguments{
\item{tappingJsonFile}{path to tapping json file}
}
\value{
data frame of tapping features
}
\description{
extracts tapping features from Tapping JSON data file
}
\examples{
library(synapseClient)
synapseLogin()
sample_Tapping_File <-'syn7067514'
tappingJsonFile <- synGet(sample_Tapping_File)@filePath
getTappingFeatures(tappingJsonFile)
}

