% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingModule.R
\name{getWalkFeatures}
\alias{getWalkFeatures}
\title{extracts walking features from walking accelerometer JSON data file}
\usage{
getWalkFeatures(walking_json_file)
}
\arguments{
\item{walking_json_file}{path to walking accelerometer json file}
}
\value{
data frame of walking features
}
\description{
extracts walking features from walking accelerometer JSON data file
}
\examples{
library(synapseClient)
synapseLogin()
sample_walking_File <-'syn7077340'
walkingJsonFile <- synGet(sample_walking_File)@filePath
getWalkFeatures(walkingJsonFile)
}
