% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gravity_data}
\alias{gravity_data}
\title{Gravity sensor measurements}
\format{An 990 x 4 data frame with column names \code{t}, \code{x},
\code{y}, \code{z} containing acceleration measurements due to gravity. 
\code{t} is the timestamp of each measurement, and \code{x}, \code{y}
and \code{z} are linear gravity measurements along the respective axis.}
\usage{
gravity_data
}
\description{
Gravity sensor measurements
}
\keyword{datasets}
