% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{guanlab_nn_architecture}
\alias{guanlab_nn_architecture}
\title{Neural net architecture for GuanLab's winning 2017 PDDB submission}
\usage{
guanlab_nn_architecture()
}
\value{
keras sequential model
}
\description{
This neural net architecture was used to generate the top 
scoring submission to the 2017 Parkinson's Disease Digital
Biomarker DREAM Challenge, Subchallenge 1
\url{https://www.synapse.org/#!Synapse:syn8717496/wiki/422884}.
}
\details{
It should be noted that GuanLab themselves wrote that:
"The parameters and structures of my network does not really matter,
as it is the simplest that can be seen in the deep learning field".
Nevertheless, this was the top scoring submission, and can be used
to generate some robust predictions to distinguish Parkinson's
patients from controls.
}
