% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guanlab_nn_weights}
\alias{guanlab_nn_weights}
\title{Weights for the included GuanLab model}
\format{A list of length 10. Each list item contains a 
matrix/array object of varying dimensionality.}
\source{
\url{https://www.synapse.org/#!Synapse:syn10146135/wiki/448409}
}
\usage{
guanlab_nn_weights
}
\description{
A list of 10 sets of weights to be used with \code{guanlab_nn_architecture}.
}
\keyword{datasets}
