% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gyroscope_data}
\alias{gyroscope_data}
\title{Gyroscope sensor measurements}
\format{
An 990 x 4 data frame with column names \code{t}, \code{x},
\code{y}, \code{z} containing accelerometer measurements. \code{t} is
the timestamp of each measurement, and \code{x}, \code{y} and \code{z}
are linear velocity measurements along the respective axis.
}
\usage{
gyroscope_data
}
\description{
Gyroscope sensor measurements
}
\keyword{datasets}
