% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{gyroscope_features_}
\alias{gyroscope_features_}
\title{Extract gyroscope features}
\usage{
gyroscope_features_(sensor_data, transform = NULL, extract = NULL,
  extract_on = c("acceleration", "velocity", "displacement"), models = NULL)
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z containing 
gyroscope measurements.}

\item{transform}{A function which accepts \code{sensor_data} as input
and outputs a dataframe.}

\item{extract}{A list of functions to be applied to each of the columns
in \code{extract_on} from the output of \code{transform}.}

\item{extract_on}{A list of column names to calculate statistics from.}
}
\value{
gyroscope features.
}
\description{
Extract gyroscope features. This function is not normally called
directly. See \code{gyroscope_features}.
}
