% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_error}
\alias{has_error}
\title{Check if a dataframe has an error column with at least one value}
\usage{
has_error(sensor_data)
}
\arguments{
\item{sensor_data}{A dataframe}
}
\value{
logical
}
\description{
Check if a dataframe has an error column with at least one value
}
