% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intertap_summary_features}
\alias{intertap_summary_features}
\title{Get default tapping features for intertap distance}
\usage{
intertap_summary_features(tapInter)
}
\arguments{
\item{tapInter}{A numeric vector containing intertap intervals}
}
\value{
A features data frame of dimension 1 x n_features
}
\description{
Calculates features characterising a timeseries data
}
