% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jerk}
\alias{jerk}
\title{Calculate jerk}
\usage{
jerk(acceleration, sampling_rate)
}
\arguments{
\item{acceleration}{An acceleration vector.}

\item{sampling_rate}{Sampling rate of the acceleration vector.}
}
\value{
Jerk vector.
}
\description{
Calculate jerk
}
