% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinetic_tremor_data}
\alias{kinetic_tremor_data}
\title{Device-Motion data from a performed kinetic tremor activity}
\format{A 3002 x 6 data frame with column names \code{attitude},
\code{timestamp}, \code{rotationRate}, \code{userAcceleration},
\code{gravity}, and \code{magneticField}. The column names correspond
to the measurements returned by a
[CMDeviceMotion](https://developer.apple.com/documentation/coremotion/cmdevicemotion)
object as described in Apple's developer documentation. \code{rotationRate},
\code{userAcceleration},\code{gravity} are each 3002 x 3 data frames with column names
\code{x}, \code{y}, \code{z} containing axial measurements for their respective sensor.
\code{magneticField} has an additional column \code{accuracy} specifying
the calibration accuracy of the magnetic field estimate. \code{attitude}
columns are quaternion coefficients.}
\usage{
kinetic_tremor_data
}
\description{
The kinetic tremor activity (also known as the finger/hand to nose test)
entails participants holding the mobile device
in their hand (either right or left) with both the mobile device and the
participants elbow at nose level. The participant then moves the mobile device
away from the nose, with both mobile device and elbow parallel to the ground
and the elbow joint acting as an axis of rotation. When the arm is fully extended,
the participant moves the mobile device back to the nose. This happens in a
continuous motion and the motion is repeated for approximately 10 seconds.
}
\keyword{datasets}
