% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{load_guanlab_model}
\alias{load_guanlab_model}
\title{Load the GuanLab model with weights.}
\usage{
load_guanlab_model()
}
\value{
list of keras sequential models
}
\description{
Load the GuanLab model with weights.
}
