% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_groups}
\alias{map_groups}
\title{Map a function to a single column within tibble groups}
\usage{
map_groups(x, col, f, ...)
}
\arguments{
\item{x}{A tibble}

\item{col}{Column to pass as a vector to \code{f}.}

\item{f}{Function to be mapped to \code{col}.}

\item{...}{Additional arguments to \code{f}.}
}
\value{
A dataframe with the original grouped columns (if \code{x} had
originally been grouped) and a \code{data} column containing
the output of the mapped function \code{f}.
}
\description{
A convenience function for mapping a function -- which accepts a
vector as input and outputs an atomic value -- to a single column
of a dataframe. If the dataframe is grouped, the function will be
applied within each group.
}
