% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_acf}
\alias{mutate_acf}
\title{Construct a dataframe with ACF values}
\usage{
mutate_acf(sensor_data, col, lag_max = NULL)
}
\arguments{
\item{sensor_data}{A data frame with column \code{col}.}

\item{col}{Name of column to calculate acf of.}

\item{lag_max}{See \code{\link[stats]{acf}}}
}
\value{
A data frame with an acf column.
}
\description{
Construct a dataframe with ACF values
}
