% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_derivative}
\alias{mutate_derivative}
\title{Add a column which is the "derivative" of an existing column to a time-series dataframe.}
\usage{
mutate_derivative(sensor_data, sampling_rate, col, derived_col)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}

\item{sampling_rate}{Sampling rate of \code{col}.}

\item{col}{Name of column to differentiate.}

\item{derived_col}{Name of new column which is the derivative of \code{col}.}
}
\value{
A dataframe
}
\description{
See function \code{derivative}.
}
