% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_detrend}
\alias{mutate_detrend}
\title{Detrend acceleration within sensor data}
\usage{
mutate_detrend(sensor_data)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}
}
\value{
Sensor data with detrended acceleration values.
}
\description{
Detrend acceleration within sensor data
}
