% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_displacement}
\alias{mutate_displacement}
\title{Add displacement column to sensor data}
\usage{
mutate_displacement(sensor_data, sampling_rate)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}

\item{sampling_rate}{Sampling rate of the acceleration data.}
}
\value{
Sensor data with displacement column.
}
\description{
Add displacement column to sensor data
}
