% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_integral}
\alias{mutate_integral}
\title{Add a column which is the "integral" of an existing column to a time-series dataframe.}
\usage{
mutate_integral(sensor_data, sampling_rate, col, derived_col)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}

\item{sampling_rate}{Sampling rate of \code{col}.}

\item{col}{Name of column to integrate.}

\item{derived_col}{Name of new column which is the integral of \code{col}.}
}
\value{
A dataframe
}
\description{
See function \code{integral}.
}
