% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_jerk}
\alias{mutate_jerk}
\title{Add jerk column to sensor data}
\usage{
mutate_jerk(sensor_data, sampling_rate)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}

\item{sampling_rate}{Sampling rate of the acceleration data.}
}
\value{
Sensor data with jerk column.
}
\description{
Add jerk column to sensor data
}
