% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_velocity}
\alias{mutate_velocity}
\title{Add velocity column to sensor data}
\usage{
mutate_velocity(sensor_data, sampling_rate)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}

\item{sampling_rate}{Sampling rate of the acceleration data.}
}
\value{
Sensor data with velocity column.
}
\description{
Add velocity column to sensor data
}
