% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{preprocess_sensor_data}
\alias{preprocess_sensor_data}
\title{Clean sensor data metrics}
\usage{
preprocess_sensor_data(sensor_data, window_length, sampling_rate,
  frequency_range, time_range)
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z containing 
sensor measurements.}

\item{window_length}{Length of sliding windows.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{time_range}{Timestamp range to use.}
}
\value{
A data frame in tidy format
}
\description{
Puts sensor data in tidy format, detrends, applies bandpass filter,
and filters time.
}
