% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{sensor_features}
\alias{sensor_features}
\title{Extract sensor features}
\usage{
sensor_features(sensor_data, transform = NULL, extract = NULL,
  extract_on = NULL, models = NULL)
}
\arguments{
\item{sensor_data}{A dataframe.}

\item{transform}{A function or list of functions to be applied sequentially
to \code{sensor_data}.}

\item{extract}{A function or list of functions to be applied to each of the
columns in \code{extract_on} from the output of \code{transform}. Each function
should return a dataframe of features (normally a single-row datafame).}

\item{extract_on}{A string or list of column names to compute features from.
If \code{NULL}, features will be extracted from all non-grouped columns.}

\item{models}{A function or list of functions which accept the output from
\code{transform} as input and output a dataframe.}
}
\value{
A list of features. The output from \code{extract} will
be stored under \code{$extracted_features} and the output from \code{models}
will be stored under \code{$model_features}. If \code{preprocess} or \code{transform}
returns an error dataframe (see \code{has_error}), the error dataframe is stored
under \code{$error}.
}
\description{
A 'pure' implementation of the feature extraction process. This function
is not normally called directly.
}
\details{
The feature extraction paradigm is implemented as such:

Input: raw sensor data
Transform: into a format suitable for calculating statistics upon
Extract: features by computing statistics upon individual columns 
         within grouped rows -- or by using a model
Return: A list of features
}
