% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_outlier_windows}
\alias{tag_outlier_windows}
\title{Identify abnormal device rotations}
\usage{
tag_outlier_windows(gravity, window_length, overlap)
}
\arguments{
\item{gravity}{A gravity vector}

\item{window_length}{Length of the filter.}

\item{overlap}{Window overlap.}
}
\value{
Rotations errors for each window.
}
\description{
Identify windows in which the phone may have been rotated or flipped,
as indicated by a gravity vector
}
