% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_outlier_windows_}
\alias{tag_outlier_windows_}
\title{Get min and max gravity values for each window}
\usage{
tag_outlier_windows_(gravity_vector, window_length, window_overlap)
}
\arguments{
\item{gravity_vector}{A gravity vector}

\item{window_length}{The number of samples per window.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap during the windowing transformation.}
}
\value{
Min and max values for each window.
}
\description{
Get min and max gravity values for each window
}
