% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tap_data}
\alias{tap_data}
\title{Sample tapping data from the tapping assay on a smartphone}
\format{A data frame with 181 rows (observations) and 4 variables:
\describe{
  \item{t}{time, in seconds}
  \item{x}{the x location of the tap on the phone screen}
  \item{y}{the y location of the tap on the phone screen}
  \item{buttonid}{a string from 'TappedButonLeft', 'TappedButtonRight' or
  'TappedButtonNone' indicating that at that time the left, right or No 
  button was tapped respectively}
}}
\usage{
tap_data
}
\description{
A dataframe containing sample JSON output format of the tap data,
containing t(time), x and y (the location of the tap on the screen),
and buttonid (which of left/right/No button was tapped).
}
\details{
Participants were shown a screen with two buttons on it, 
indicating left and right. They were asked to tap on the buttons
alternatingly as fast as they can for 30s.
}
\keyword{datasets}
