% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tapdrift_summary_features}
\alias{tapdrift_summary_features}
\title{Get default tapping features for tap drift}
\usage{
tapdrift_summary_features(tapDrift)
}
\arguments{
\item{tapDrift}{A numeric vector}
}
\value{
A features data frame of dimension 1 x n_features
}
\description{
Calculates features characterising a timeseries data
}
