% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_sensor_data}
\alias{tidy_sensor_data}
\title{Gather the axial columns}
\usage{
tidy_sensor_data(sensor_data)
}
\arguments{
\item{sensor_data}{A data frame with a time column \code{t},
and at least one additional axial column.}
}
\value{
Sensor data in tidy format.
}
\description{
Gather x, y, and z columns into a single \code{axis} column and
normalize the \code{t} column to begin at \code{t} = 0.
}
