% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction_functions.R
\name{time_domain_summary}
\alias{time_domain_summary}
\title{Returns statistical summary of the time series}
\usage{
time_domain_summary(values, sampling_rate = NA)
}
\arguments{
\item{values}{A numeric vector.}

\item{sampling_rate}{Sampling_rate of \code{values}. If NA it uses default
sampling rate of 100Hz.}
}
\value{
A features data frame of dimension 1 x 20. See the
feature definition vignette:
\code{vignette("feature_definitions", package="mhealthtools")}
}
\description{
A convenience feature extraction function which characterises a given
time series in to statistical features in the time domain.
}
\examples{
time_features = time_domain_summary(
  accelerometer_data$x,
  sampling_rate = 100.122)
}
\author{
Thanneer Malai Perumal, Phil Snyder
}
