% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{transform_accelerometer_data}
\alias{transform_accelerometer_data}
\title{Prepare accelerometer sensor data for feature extraction}
\usage{
transform_accelerometer_data(sensor_data, transformation = NULL,
  window_length = 256, time_range = c(1, 9), frequency_range = c(1, 25),
  sampling_rate = 100)
}
\arguments{
\item{sensor_data}{A dataframe with columns t, x, y, z.}

\item{transformation}{A function which accepts a tidy version of \code{sensor_data},
with columns t, axis, value, as input and outputs a dataframe suitable for 
feature extraction.}

\item{window_length}{Length of sliding windows during bandpass filtering.}

\item{time_range}{Timestamp range to use.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{sampling_rate}{Sampling rate of the acceleration column.}
}
\value{
A dataframe
}
\description{
Prepare accelerometer sensor data for feature extraction
}
