% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{transform_gyroscope_data}
\alias{transform_gyroscope_data}
\title{Prepare gyroscope sensor data for feature extraction}
\usage{
transform_gyroscope_data(sensor_data, transformation = NA,
  window_length = 256, overlap = 0.5, time_range = c(1, 9),
  frequency_range = c(1, 25), sampling_rate = 100, groups = c("axis",
  "Window"))
}
\arguments{
\item{sensor_data}{A dataframe with columns t, x, y, z.}

\item{transformation}{A function which accepts a tidy version of \code{sensor_data},
with columns t, axis, value, as input and outputs a dataframe suitable for 
feature extraction. By default, the tidy sensor data is windowed before feature
extraction. This is a more user-friendly version of the \code{transform} 
parameter required for more generic functions like \code{sensor_features}. Namely,
its input will already be "cleaned" by detrending, bandpassing, etc. See 
\code{preprocess_sensor_data} for all the cleaning steps performed.}

\item{window_length}{Length of sliding windows.}

\item{overlap}{Window overlap.}

\item{time_range}{Timestamp range to use.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{sampling_rate}{Sampling rate of the acceleration vector.}

\item{groups}{A list of column names to group upon when calculating statistics}
}
\value{
A dataframe
}
\description{
Prepare gyroscope sensor data for feature extraction
}
