% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{transform_kinematic_sensor_data}
\alias{transform_kinematic_sensor_data}
\title{Apply standard transformations to kinematic sensor data}
\usage{
transform_kinematic_sensor_data(sensor_data, transformation, window_length,
  time_range, frequency_range, sampling_rate)
}
\arguments{
\item{sensor_data}{A dataframe with columns t, x, y, z.}

\item{transformation}{A function which accepts a tidy version of \code{sensor_data},
with columns t, axis, value, as input and outputs a dataframe suitable for 
feature extraction. By default, the tidy sensor data is windowed before feature
extraction. This is a more user-friendly version of the \code{transform} 
parameter required for more generic functions like \code{sensor_features}. Namely,
its input will already be "cleaned" by detrending, bandpassing, etc. See 
\code{preprocess_sensor_data} for all the cleaning steps performed.}

\item{window_length}{Length of sliding windows during bandpass filtering.}

\item{time_range}{Timestamp range to use.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{sampling_rate}{Sampling rate of the acceleration vector.}
}
\value{
A dataframe.
}
\description{
Apply standard transformations to kinematic (accelerometer/gyroscope)
sensor data. This function is not normally called directly. See
\code{transform_accelerometer_data} and \code{transform_gyroscope_data}.
}
