% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{transformation_imf_window}
\alias{transformation_imf_window}
\title{Generate a function for windowing sensor data after applying EMD}
\usage{
transformation_imf_window(window_length, overlap, max_imf)
}
\arguments{
\item{window_length}{Length of the sliding window}

\item{overlap}{Window overlap}

\item{max_imf}{The maximum number of IMF's during EMD}
}
\value{
A function that accepts as input a dataframe with columns 
t, axis, value and outputs a windowed transformation of that 
dataframe's EMD.
}
\description{
Apply empirical mode decomposition to sensor data, then window the result.
}
