% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{transformation_window}
\alias{transformation_window}
\title{Generate a function for applying a window transformation to sensor data}
\usage{
transformation_window(window_length, window_overlap)
}
\arguments{
\item{window_length}{Length of the sliding window.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap.}
}
\value{
A function that accepts as input a dataframe with columns
t, axis, value and outputs a windowed transformation of that dataframe
}
\description{
Generate a function for applying a window transformation to sensor data
}
