% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{velocity}
\alias{velocity}
\title{Calculate velocity}
\usage{
velocity(acceleration, sampling_rate)
}
\arguments{
\item{acceleration}{An acceleration vector.}

\item{sampling_rate}{Sampling rate of the acceleration vector.}
}
\value{
Velocity vector.
}
\description{
Calculate velocity
}
