% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{window}
\alias{window}
\title{Window the acceleration vector of sensor data by axis}
\usage{
window(sensor_data, window_length, overlap)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, acceleration.}

\item{window_length}{Length of the filter}

\item{overlap}{window overlap}
}
\value{
Windowed sensor data
}
\description{
Window the acceleration vector of sensor data by axis
}
