% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{window_signal}
\alias{window_signal}
\title{Window a signal}
\usage{
window_signal(values, window_length = 256, window_overlap = 0.5,
  window_name = "hamming")
}
\arguments{
\item{values}{Timeseries vector of length n.}

\item{window_length}{The number of samples per window.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap.}

\item{window_name}{window name: bartlett, blackman, flattop, hamming,
hanning, or rectangle. See \code{\link[seewave]{ftwindow}}.}
}
\value{
A matrix of window_length x nwindows
}
\description{
Given a numeric vector, this function will return its windowed signal.
}
