% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{window_start_end_times}
\alias{window_start_end_times}
\title{Compute start/end timestamps for each window}
\usage{
window_start_end_times(t, window_length, window_overlap)
}
\arguments{
\item{t}{A numeric time vector}

\item{window_length}{The number of samples per window.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap.}
}
\value{
A dataframe with columns window, window_start_time,
window_end_time, window_start_index, window_end_index
}
\description{
Compute start/end timestamps for each window
}
