% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{accelerometer_features}
\alias{accelerometer_features}
\title{Extract accelerometer features}
\usage{
accelerometer_features(sensor_data, funs = c(time_domain_summary,
  frequency_domain_summary, frequency_domain_energy), window_length = 256,
  overlap = 0.5, time_range = c(1, 9), frequency_range = c(1, 25))
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z containing 
accelerometer measurements.}

\item{funs}{Feature extraction functions that accept a single
time-series vector as input.}

\item{window_length}{Length of sliding windows.}

\item{overlap}{Window overlap.}

\item{time_range}{Timestamp range to use.}

\item{frequency_range}{Frequency range for the bandpass filter.}
}
\value{
Accelerometer features.
}
\description{
Extract features summarizing time domain, frequency domain, and frequency energy
from accelerometer measurements.
}
