% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_functions.R, R/sensors.R
\name{accelerometer_features}
\alias{accelerometer_features}
\title{Extract accelerometer features}
\usage{
accelerometer_features(sensor_data, time_filter = NULL, detrend = F,
  frequency_filter = NULL, IMF = 1, window_length = NULL,
  window_overlap = NULL, derived_kinematics = F, funs = NULL,
  models = NULL)

accelerometer_features(sensor_data, time_filter = NULL, detrend = F,
  frequency_filter = NULL, IMF = 1, window_length = NULL,
  window_overlap = NULL, derived_kinematics = F, funs = NULL,
  models = NULL)
}
\arguments{
\item{sensor_data}{A \code{n} x 4 data frame with column names \code{t}, \code{x},
\code{y}, \code{z} containing accelerometer measurements. Here \code{n} is the
total number of measurements, \code{t} is the time of measurement,
\code{x}, \code{y} and \code{z} are the linear acceleration measurements excluding 
gravity in their respective coordinates of the smartphone tri-axial accelerometers. 
x, y, and z co-ordinates are with respect to phones axis as reference}

\item{time_filter}{A length 2 numeric vector specifying the time range to use 
during preprocessing and feature extraction after normalizing the first timestamp
to zero. A \code{NULL} value means do not filter time.}

\item{detrend}{Whether to detrend the signal. A \code{NULL} value means do not detrend.}

\item{frequency_filter}{A length 2 numeric vector specifying the frequency range
of the signal to use during preprocessing and feature extraction. A \code{NULL} 
value means do not filter frequencies.}

\item{IMF}{The number of IMFs used during empirical mode decomposition. 
The default value 1 means do not apply EMD to the signal.}

\item{window_length}{Length of the sliding window used during a windowing 
transformation. Both \code{window_length} and \code{window_overlap} must be
set for the windowing transformation to be applied.}

\item{window_overlap}{Window overlap used during a windowing transformation.
Both \code{window_length} and \code{window_overlap} must be set for the
transformation to be applied.}

\item{derived_kinematics}{Whether to add columns for \code{jerk}, \code{velocity},
and \code{displacement} before extracting features.}

\item{funs}{A list of feature extraction functions that each accept
a single numeric vector as input. Each function should return a 
dataframe of features (normally a single-row datafame). The input vectors
will be the axial measurements from \code{sensor_data} after the chosen
preprocessing and transformation steps have been applied. If no argument
is supplied to either \code{funs} or \code{models}, a default set
of feature extraction functions (as described in \code{default_kinematic_features})
will be supplied for this parameter.}

\item{models}{A list of functions, each of which accept as input 
\code{sensor_data} after the chosen preprocessing and transformation
steps have been applied and return features. Useful for models which compute
individual statistics using multiple input variables.}

\item{mutate}{Whether to add columns for \code{jerk}, \code{velocity},
and \code{displacement} before extracting features.}

\item{sensor_data}{An \code{n} x 4 data frame with column names \code{t}, \code{x},
\code{y}, \code{z} containing accelerometer measurements. Here \code{n} is the
total number of measurements, \code{t} is the timestamp of each measurement, and
\code{x}, \code{y} and \code{z} are linear acceleration measurements.}

\item{time_filter}{A length 2 numeric vector specifying the time range 
of measurements to use during preprocessing and feature extraction after
normalizing the first timestamp to 0. A \code{NULL} value means do not 
filter any measurements.}

\item{detrend}{Whether to detrend the signal.}

\item{frequency_filter}{A length 2 numeric vector specifying the frequency range
of the signal (in hertz) to use during preprocessing and feature extraction.
A \code{NULL} value means do not filter frequencies.}

\item{IMF}{The number of IMFs used during an empirical mode decomposition
transformation. The default value of 1 means do not apply EMD to the signal.}

\item{window_length}{Length of the sliding window used during the windowing 
transformation. Both \code{window_length} and \code{window_overlap} must be
set for the windowing transformation to be applied.}

\item{window_overlap}{Window overlap used during the windowing transformation.
Both \code{window_length} and \code{window_overlap} must be set for the
windowing transformation to be applied.}

\item{funs}{A function or list of feature extraction functions that each
accept a single numeric vector as input. Each function should return a 
dataframe of features (normally a single-row datafame). The input vectors
will be the axial measurements from \code{sensor_data} after the chosen
preprocessing and transformation steps have been applied. If no argument
is supplied to either \code{funs} or \code{models}, a default set
of feature extraction functions (as described in \code{default_kinematic_features})
will be supplied for this parameter.}

\item{models}{A list of functions, each of which accept as input 
\code{sensor_data} after the chosen preprocessing and transformation
steps have been applied and return features. Useful for models which compute
individual statistics using multiple input variables.}
}
\value{
A list of accelerometer features. The output from \code{funs} will
be stored under \code{$extracted_features} and the output from \code{models}
will be stored under \code{$model_features}. If there is an error during
extraction, the returned result will be stored under \code{$error}.
}
\description{
Extract accelerometer features

Extract accelerometer features
}
