% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{accelerometer_features_}
\alias{accelerometer_features_}
\title{Extract accelerometer features}
\usage{
accelerometer_features_(sensor_data, transform = NULL, extract = NULL,
  extract_on = c("acceleration", "jerk", "velocity", "displacement"),
  models = NULL)
}
\arguments{
\item{sensor_data}{A data frame with columns t, x, y, z containing 
accelerometer measurements.}

\item{transform}{A function which accepts \code{sensor_data} as input
and outputs a dataframe.}

\item{extract}{A list of functions to be applied to each of the columns
in \code{extract_on} from the output of \code{transform}. Each function
should return a dataframe of features (normally a single-row datafame).}

\item{extract_on}{A list of column names to calculate statistics from.}

\item{extract_on}{A list of column names to calculate statistics from.}
}
\value{
A list of accelerometer features. The output from \code{extract} will
be stored under \code{$extracted_features} and the output from \code{models}
will be stored under \code{$model_features}.
}
\description{
Extract accelerometer features. This function is not normally called
directly. See \code{accelerometer_features}.
}
