% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bandpass}
\alias{bandpass}
\title{Apply a pass-band filter to time series data}
\usage{
bandpass(values, window_length, sampling_rate, frequency_range)
}
\arguments{
\item{values}{Numeric vector.}

\item{window_length}{Length of the filter.}

\item{sampling_rate}{Sampling rate of the values.}

\item{frequency_low}{Lower bound on frequency in Hz}

\item{frequency_high}{Upper bound on frequency in Hz}
}
\value{
Filtered time series data
}
\description{
Apply a pass-band filter to time series data
}
