% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_acf}
\alias{calculate_acf}
\title{Construct a dataframe with ACF values}
\usage{
calculate_acf(sensor_data, col)
}
\arguments{
\item{sensor_data}{A data frame with columns \code{axis}, \code{Window}, and \code{col}.}

\item{col}{Name of column to calculate acf of.}
}
\value{
A tibble with columns axis, Window, window_index, acf
}
\description{
Estimate the ACF for windowed sensor data.
}
