% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{default_kinematic_features}
\alias{default_kinematic_features}
\title{Get list of default feature extraction functions}
\usage{
default_kinematic_features(sampling_rate)
}
\arguments{
\item{sampling_rate}{Sampling rate of the data these functions will be
applied to.}
}
\value{
A list of closures for three feature extraction functions supplied
with this package that can be passed to the \code{extract} parameter
in \code{*_features} functions.
}
\description{
Get list of default feature extraction functions
}
