% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detrend}
\alias{detrend}
\title{Detrend time series data}
\usage{
detrend(time, values)
}
\arguments{
\item{time}{Numeric vector containing the timestamp values.}

\item{values}{Numeric vector the same length as \code{time}.}
}
\value{
Numeric vector with detrended values.
}
\description{
Detrend time series data
}
