% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_features}
\alias{extract_features}
\title{Extract features from a column}
\usage{
extract_features(x, col, funs)
}
\arguments{
\item{x}{A data frame with columns axis, window, and \code{col}.}

\item{col}{The name of the column in \code{x} to pass to each 
function in \code{funs}.}

\item{funs}{A list of functions that accept a single vector as input.}
}
\value{
a data frame with columns axis, window, and other feature columns.
}
\description{
Apply each of the functions in \code{funs} to the column 
\code{col} in data frame \code{x}. Each of the functions in 
\code{funs} must accept a single vector as input and output 
a data frame with columns axis and window (and optionally others).
}
