% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_time}
\alias{filter_time}
\title{Select a specific time range from sensor data.}
\usage{
filter_time(sensor_data, t1, t2)
}
\arguments{
\item{t1}{Start time.}

\item{t2}{End time.}

\item{sensor_date}{A data frame with a \code{t} column.}
}
\value{
Sensor data between time t1 and t2 (inclusive)
}
\description{
Select a specific time range from sensor data.
}
