% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction_functions.R
\name{frequency_domain_energy}
\alias{frequency_domain_energy}
\title{Returns energy for each 0.5Hz band in the frequency spectrum.}
\usage{
frequency_domain_energy(values, sampling_rate = NA)
}
\arguments{
\item{values}{A numeric vector.}

\item{sampling_rate}{Sampling_rate of \code{values}. If NA it uses default
sampling rate of 100Hz.}
}
\value{
A features data frame of dimension 1 x 48, each representing energy within
a 0.5Hz band. See the
feature definition vignette:
\code{vignette("feature_definitions", package="mhealthtools")}
}
\description{
A convenience feature extraction function that converts a given
time series in to frequency spectrum and computes energy in each
0.5Hz band from 0 to 25 Hz.
}
\examples{
frequency_energy_features = frequency_domain_energy(
  accelerometer_data$x,
  sampling_rate = 100.122)
}
\author{
Thanneer Malai Perumal, Phil Snyder
}
