% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frequency_domain_energy}
\alias{frequency_domain_energy}
\title{Get frequency domain energy features}
\usage{
frequency_domain_energy(values, sampling_rate = NA)
}
\arguments{
\item{values}{A timeseries vector.}

\item{sampling_rate}{Sampling rate of the signal (by default it is 100 Hz).}
}
\value{
A features data frame of dimension 1 x 48.
}
\description{
Given a numeric vector, this function will return features
characterising the time series in frequency domain.
}
