% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frequency_domain_summary}
\alias{frequency_domain_summary}
\title{Get frequency domain features}
\usage{
frequency_domain_summary(values, sampling_rate = NA, npeaks = NA)
}
\arguments{
\item{values}{A numeric vector.}

\item{sampling_rate}{Sampling_rate of \code{values}.}

\item{npeaks}{Number of peaks to be computed in EWT}
}
\value{
A features data frame of dimension 1 x num_features
}
\description{
Calculates features characterising a time series in the frequency domain.
}
