% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{get_filtered_signal}
\alias{get_filtered_signal}
\title{Bandpass and sorted mean filter the given signal}
\usage{
get_filtered_signal(x, sampling_rate, mean_filter_order = 65, method = "acf")
}
\arguments{
\item{x}{A time series numeric data}

\item{sampling_rate}{The sampling rate (fs) of the signal}

\item{mean_filter_order}{The number of samples used in the sliding window
for the mean filtering function}

\item{method}{The algorithm used to estimate the heartrate, because the
preprocessing steps are different for each. method can be any of
'acf','psd' or 'peak' for algorithms based on autocorrelation,
power spectral density and peak picking respectively}
}
\value{
The filtered time series data
}
\description{
Bandpass and sorted mean filter the given signal
}
