% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{get_filtered_signal}
\alias{get_filtered_signal}
\title{Bandpass and sorted mean filter the given signal}
\usage{
get_filtered_signal(x, sampling_rate, mean_filter_order = 65,
  method = "acf")
}
\arguments{
\item{x}{A time series numeric data}
}
\value{
The filtered time series data
}
\description{
Bandpass and sorted mean filter the given signal
}
