% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{get_heartrate}
\alias{get_heartrate}
\title{Preprocess and extract heart rate from smartphone video recordings.}
\usage{
get_heartrate(heartrate_data, window_length = 10, window_overlap = 0.5)
}
\arguments{
\item{heartrate_data}{A data frame with columns t, red, green and blue}

\item{window_length}{Length of the time window in seconds, to be considered
while calculating the heart rate for each channel}

\item{frequency_range}{Frequency range in Hz for the bandpass filter parameters}

\item{bandpass_order}{Order (length) of the bandpass filter to be used for filtering}
}
\value{
list containing heart rate and confidence of the estimate for
each color (red, green, blue)
}
\description{
A convenience wrapper for extracting heart rate features for each color
band from average pixel value per frame of video (processed hr) captured
using smartphone cameras.
}
\examples{
heartrate_data = heartrate_data[,c('timestamp', 'red', 'green', 'blue')]
heartrate_ftrs = get_heartrate(heartrate_data)
 
}
\author{
Meghasyam Tummalacherla, Phil Snyder
}
