% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heartrate.R
\name{get_heartrate}
\alias{get_heartrate}
\title{Preprocess and extract heart rate from smartphone video recordings.}
\usage{
get_heartrate(heartrate_data, window_length = 10, window_overlap = 0.5,
  method = "acf")
}
\arguments{
\item{heartrate_data}{A data frame with columns t, red, green and blue}

\item{window_length}{Length of the time window \emph{in seconds}, to be
considered while calculating the heart rate for each channel.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap.}

\item{method}{The algorithm used to calculate the heartrate, current methods
include ('acf','psd') which stand for autocorrelation function, and power
spectral density respectively. We will be adding support for peak picking 
algorithms, and wavelet methods later.}
}
\value{
list containing heart rate and confidence of the estimate for
each color (red, green, blue)
}
\description{
A convenience wrapper for extracting heart rate features for each color
band from average pixel value per frame of video (processed hr) captured
using smartphone cameras.
}
\examples{
heartrate_data = heartrate_data[,c('t', 'red', 'green', 'blue')]
heartrate_ftrs = get_heartrate(heartrate_data)
 
}
\author{
Meghasyam Tummalacherla, Phil Snyder
}
