% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinetic_tremor_features.R
\name{get_kinetic_tremor_features}
\alias{get_kinetic_tremor_features}
\title{extracts features from tremor task handToNose accelerometer and gyroscope JSON data file}
\usage{
get_kinetic_tremor_features(accelerometer_data, gyroscope_data,
  gravity_data = NA, funs = NA, window_length = 256, time_range = c(1,
  9), frequency_range = c(1, 25), overlap = 0.5, max_imf = 4)
}
\arguments{
\item{tremorJsonFileLoc}{path to tremor accelerometer json file}
}
\value{
data frame of tremor features
}
\description{
extracts features from tremor task handToNose accelerometer and gyroscope JSON data file
}
\author{
Thanneer Malai Perumal, Meghasyam Tummalacherla, Phil Snyder
}
