% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_walk_features.R
\name{get_walk_features}
\alias{get_walk_features}
\title{Extract walk features from raw accelerometer and gyroscope data.}
\usage{
get_walk_features(accelerometer_data = NULL, gyroscope_data = NULL,
  gravity_data = NULL, time_filter = NULL, detrend = F,
  frequency_filter = NULL, IMF = 1, window_length = NULL,
  window_overlap = NULL, derived_kinematics = F, funs = NULL,
  models = NULL)
}
\arguments{
\item{accelerometer_data}{A data frame with columns t, x, y, z containing 
accelerometer measurements.}

\item{gyroscope_data}{A data frame with columns t, x, y, z containing 
gyroscope measurements.}

\item{gravity_data}{A data frame with columns t, x, y, z containing 
gravity sensor measurements.}

\item{window_length}{Length of sliding windows for bandpass filter
and windowing transformation.}

\item{funs}{Feature extraction functions that accept a single
time-series vector as input.}

\item{models}{A function which accepts as input a dataframe with columns
axis, window, IMF, jerk, acceleration, velocity, displacement and
outputs features. Useful for models which compute individual statistics
using multiple input variables.}

\item{time_range}{Timestamp range to use.}

\item{frequency_range}{Frequency range for the bandpass filter.}

\item{overlap}{Window overlap for the windowing transformation.}

\item{max_imf}{Number of intrinsic mode functions to use for 
empirical mode decomposition.}
}
\value{
A list of feature dataframes. The outputs from \code{funs} will
be stored under \code{$extracted_features} and the outputs from \code{models}
will be stored under \code{$model_features}.'
}
\description{
Extract walk features from raw accelerometer and gyroscope data.
}
\author{
Thanneer Malai Perumal, Meghasyam Tummalacherla, Phil Snyder
}
