% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{GetLeftRightEventsAndTapIntervals}
\alias{GetLeftRightEventsAndTapIntervals}
\title{Curate the raw tapping data to get Left and Right events, after applying the threshold}
\usage{
GetLeftRightEventsAndTapIntervals(tapData, depressThr = 20)
}
\arguments{
\item{error_message}{A text error message which would be reflected in the feature output}
}
\value{
A dataframe with NAs as feature values and the appropriate error message
}
\description{
Curate the raw tapping data to get Left and Right events, after applying the threshold
}
