% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restModule.R
\name{getRestFeatures}
\alias{getRestFeatures}
\title{extracts features from accelerometer rest data file}
\usage{
getRestFeatures(restAccel_json_file)
}
\arguments{
\item{restAccel_json_file}{path to accelerometer json file}
}
\value{
data frame of rest features
}
\description{
extracts features from accelerometer rest data file
}
\examples{
library(synapseClient)
synapseLogin()
walkingTable <- synTableQuery("SELECT * FROM syn5713119")
walkingTable <- walkingTable@values
sampleRow <- rownames(walkingTable)[10]
sample_restAccel_jsonFile <- synDownloadTableFile('syn5713119',sampleRow, 'deviceMotion_walking_rest.json.items')
sample_restAccel_jsonFile <- as.character(sample_restAccel_jsonFile)
sample_restResults <- getRestFeatures(sample_restAccel_jsonFile)
}
