% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heartrate_data}
\alias{heartrate_data}
\title{Sample heartrate data from a smartphone camera}
\format{A data frame with 3661 rows (observations) and 7 variables:
\describe{
  \item{t}{time, in seconds}
  \item{green}{mean green intensity across the frame at the given timestamp}
  \item{blue}{mean blue intensity across the frame at the given timestamp}
  \item{red}{mean red intensity across the frame at the given timestamp}
  \item{brightness}{mean brightness across the frame at the given timestamp}
  \item{hue}{mean hue across the frame at the given timestamp}
  \item{saturation}{mean saturation across the frame at the given timestamp}
}}
\usage{
heartrate_data
}
\description{
A dataframe containing sample JSON output format of the heartrate data,
containing red, green, blue levels from the camera sensor, along with the
timestamp, and the derived metrics of hue, saturation and brightness.
}
\details{
A video(240p) of a finger covering the smartphone camera with the flash
turned on was recorded at 60fps. The average intensity of all the pixels per
frame in all the three color channels was calculated, and indexed along
the timestamp of the frame. We also have hue, brightness and saturation
calculated in a similar way.
}
\keyword{datasets}
