% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{kinematic_sensor_features}
\alias{kinematic_sensor_features}
\title{Extract kinematic sensor features}
\usage{
kinematic_sensor_features(sensor_data, transform, extract, extract_on, groups,
  acf_col)
}
\arguments{
\item{sensor_data}{A dataframe.}

\item{transform}{A function which accepts \code{sensor_data} as input
and outputs a dataframe.}

\item{extract}{A list of functions to be applied to each of the columns
in \code{extract_on} from the output of \code{transform}.}

\item{extract_on}{A list of column names to calculate statistics from.}

\item{groups}{A list of column names to group upon when calculating statistics.}

\item{acf_col}{Column name to calculate acf upon.}
}
\value{
A dataframe of features.
}
\description{
Extract kinematic (accelerometer/gyroscope) features. This function is not 
normally called directly. See \code{accelerometer_features} and \code{gyroscope_features}.
}
