% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_groups}
\alias{map_groups}
\title{Map a function to a single column within tibble groups}
\usage{
map_groups(x, col, f, ...)
}
\arguments{
\item{col}{Column to pass as a vector to \code{.f}.}

\item{...}{Additional arguments to \code{.f}.}

\item{.x}{A tibble}

\item{.f}{Function to be mapped to \code{col} for each group.}
}
\value{
A tibble indexed by groups with an additional column containing
the output of the mapped function.
}
\description{
A convenience function for mapping a function -- which accepts a vector as input
and outputs an atomic value -- to a single column of each group in a grouped tibble.
}
