% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_bandpass}
\alias{mutate_bandpass}
\title{Apply a pass-band filter to sensor data}
\usage{
mutate_bandpass(sensor_data, window_length, sampling_rate, frequency_range)
}
\arguments{
\item{sensor_data}{A data frame with column value.}

\item{window_length}{Length of the filter.}

\item{sampling_rate}{Sampling rate of the value column.}

\item{frequency_range}{Bounds on frequency in Hz}
}
\value{
Filtered time series data
}
\description{
Apply a pass-band filter to sensor data
}
