% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_detrend}
\alias{mutate_detrend}
\title{Detrend sensor data}
\usage{
mutate_detrend(sensor_data)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, value.}
}
\value{
Sensor data with detrended values.
}
\description{
Detrend sensor data
}
