% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{prepare_kinematic_sensor_args}
\alias{prepare_kinematic_sensor_args}
\title{Return arguments to be used in general feature functions}
\usage{
prepare_kinematic_sensor_args(sensor_data, metric, time_filter = NULL,
  detrend = F, frequency_filter = NULL, IMF = 1,
  window_length = NULL, window_overlap = NULL,
  derived_kinematics = F, funs = NULL, models = NULL)
}
\arguments{
\item{sensor_data}{An \code{n} x 4 data frame with column names \code{t}, \code{x},
\code{y}, \code{z} containing gyroscope measurements. Here \code{n} is the
total number of measurements, \code{t} is the timestamp of each measurement, and
\code{x}, \code{y} and \code{z} are linear velocity measurements.}

\item{metric}{Name of the metric measured by this sensor. For accelerometer
data, the metric is acceleration. Whereas for gyroscope data the metric is
velocity.}

\item{time_filter}{A length 2 numeric vector specifying the time range 
of measurements to use during preprocessing and feature extraction after
normalizing the first timestamp to 0. A \code{NULL} value means do not 
filter any measurements.}

\item{detrend}{Whether to detrend the signal.}

\item{frequency_filter}{A length 2 numeric vector specifying the frequency range
of the signal (in hertz) to use during preprocessing and feature extraction.
A \code{NULL} value means do not filter frequencies.}

\item{IMF}{The number of IMFs used during an empirical mode decomposition
transformation. The default value of 1 means do not apply EMD to the signal.}

\item{window_length}{Length of the sliding window used during the windowing 
transformation. Both \code{window_length} and \code{window_overlap} must be
set for the windowing transformation to be applied.}

\item{window_overlap}{Window overlap used during the windowing transformation.
Both \code{window_length} and \code{window_overlap} must be set for the
windowing transformation to be applied.}

\item{derived_kinematics}{Whether to add columns for \code{jerk}, \code{velocity},
and \code{displacement} before extracting features.}

\item{funs}{A function or list of feature extraction functions that each
accept a single numeric vector as input. Each function should return a 
dataframe of features (normally a single-row datafame). The input vectors
will be the axial measurements from \code{sensor_data} after the transform
defined by the above parameters has been applied. If no argument
is supplied to either \code{funs} or \code{models}, a default set
of feature extraction functions (as described in \code{default_kinematic_features})
will be supplied for this parameter.}

\item{models}{A list of functions, each of which accepts
\code{sensor_data} as input after the transform defined by the above 
parameters has been applied and returns features. Useful for models
which compute individual features using multiple input variables.}
}
\value{
A list of arguments to be used in the general feature functions.
}
\description{
Using the arguments passed to convenience functions like 
\code{accelerometer_features} and \code{gyroscope_features},
build an argument set that can be used with more general functions
like \code{kinematic_sensor_data} and \code{sensor_data}. This function
is not normally called directly. See \code{accelerometer_features} and
\code{gyroscope_features}.
}
