% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{sensor_features}
\alias{sensor_features}
\title{Extract features from sensor data}
\usage{
sensor_features(sensor_data, transform, extract, extract_on, groups)
}
\arguments{
\item{sensor_data}{A dataframe.}

\item{transform}{A function which accepts \code{sensor_data} as input
and outputs a dataframe.}

\item{extract}{A list of functions to be applied to each of the columns
in \code{extract_on} from the output of \code{transform}.}

\item{extract_on}{A list of column names to calculate statistics from.}

\item{groups}{A list of column names to group upon when calculating statistics.}
}
\value{
A dataframe of features
}
\description{
A 'pure' implementation of the feature extraction process. This function
is not normally called directly.
}
\details{
The feature extraction paradigm is implemented as such:

Input: raw feature data
Transform: into a format suitable for calculating statistics upon
Extract: features by calculating statistics upon individual, grouped columns
Return: statistics/features for each group
}
