% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_outlier_windows}
\alias{tag_outlier_windows}
\title{Identify abnormal device rotations}
\usage{
tag_outlier_windows(gravity, window_length, window_overlap)
}
\arguments{
\item{gravity}{A dataframe with gravity vectors for columns}

\item{window_length}{The number of samples per window.}

\item{window_overlap}{Fraction in the interval [0, 1) specifying the amount of
window overlap during the windowing transformation.}
}
\value{
Rotations errors for each window.
}
\description{
Identify windows in which the phone may have been rotated or flipped,
as indicated by a gravity vector
}
