% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_outlier_windows_}
\alias{tag_outlier_windows_}
\title{Get min and max gravity values for each window}
\usage{
tag_outlier_windows_(gravity_vector, window_length, overlap)
}
\arguments{
\item{gravity_vector}{A gravity vector}

\item{window_length}{Length of the filter.}

\item{overlap}{Window overlap.}
}
\value{
Min and max values for each window.
}
\description{
Get min and max gravity values for each window
}
