% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{tapping_features}
\alias{tapping_features}
\title{Extract tapping (screen sensor) features}
\usage{
tapping_features(tap_data, depressThr = 20)
}
\arguments{
\item{tap_data}{A data frame with columns t, x, y, buttonid containing 
tapping measurements. buttonid can be from c('TappedButtonLeft','TappedButtonRight','TappedButtonNone') 
indicating a tap that has been classified as to the left, right or neither of those places on the screen}

\item{depressThr}{A numerical threshold for intertap distance in x axis}
}
\value{
A dataframe of features.
}
\description{
Extract tapping (screen sensor) features
}
