% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_domain_summary}
\alias{time_domain_summary}
\title{Get time domain features}
\usage{
time_domain_summary(accel, sampling_rate = NA)
}
\arguments{
\item{accel}{An acceleration vector.}

\item{sampling_rate}{Sampling_rate of the acceleration vector.}
}
\value{
A features data frame of dimension 1 x n_features
}
\description{
Calculates features characterising a time series in the time domain.
}
