% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtractionFunctions.R
\name{time_domain_summary}
\alias{time_domain_summary}
\title{Returns statistical summary of the time series}
\usage{
time_domain_summary(values, samplig_rate)
}
\arguments{
\item{values}{A numeric vector.}

\item{sampling_rate}{Sampling_rate of \code{values}. If NA it uses default 
sampling rate of 100Hz.}
}
\value{
A features data frame of dimension 1 x 20. See 
https://github.com/Sage-Bionetworks/mhealthtools/blob/master/FeatureDefinitions.md
for feature definitions.
}
\description{
\code{time_domain_summary()} is a convinient feature 
extraction function that characterises a given time series in to 
statistical features in the time domain.
}
\examples{
library(mhealthtools)

data(accelerometer_data)

tm.ftrs = time_domain_summary(accelerometer_data$x, sampling_rate = 100.122)

}
\author{
Thanneer Malai Perumal, Phil Snyder
}
