% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{window}
\alias{window}
\title{Window the value vector of sensor data for each axis}
\usage{
window(sensor_data, window_length, overlap)
}
\arguments{
\item{sensor_data}{A data frame with columns t, axis, value.}

\item{window_length}{Length of the filter}

\item{overlap}{window overlap}

\item{include_timestamp}{Whether to include columns for starting and ending
timestamps for each row.}
}
\value{
Windowed sensor data
}
\description{
Window the value vector of sensor data for each axis
}
