% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{window_start_end_times}
\alias{window_start_end_times}
\title{Compute start/end timestamps for each window}
\usage{
window_start_end_times(t, window_length, overlap)
}
\arguments{
\item{t}{A numeric time vector}

\item{window_length}{Length of the filter}

\item{overlap}{Window overlap}
}
\value{
A dataframe with columns Window, window_start_time,
window_end_time, window_start_index, window_end_index
}
\description{
Compute start/end timestamps for each window
}
