% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{windowSignal}
\alias{windowSignal}
\title{Window a signal
 
Given a numeric vector, this function will return a windowed 
signal with hamming window.}
\usage{
windowSignal(values, window_length = 256, overlap = 0.5)
}
\arguments{
\item{values}{Timeseries vector of length n.}

\item{window_length}{Length of the filter.}

\item{overlap}{Window overlap.}
}
\value{
A matrix of window_length x nwindows
}
\description{
Window a signal
 
Given a numeric vector, this function will return a windowed 
signal with hamming window.
}
